dag file column type for hive table changed

+++ dag_file.py
@@ -55,7 +55,7 @@ with DAG("ecommerce_platform",start_date=datetime(2021, 1, 1),
                     `start` TIMESTAMP,
                     `end` TIMESTAMP,
                     source STRING,
-                    source_number INTEGER
+                    source_number LONG
                     )
                 ROW FORMAT DELIMITED
                 FIELDS TERMINATED BY ','
